test: add unit tests for i18n path helpers
